Remove flatMapMutable
This commit removes the lazy flatMapMutable function as it seems to not be possible to implement correctly with a sequence-numbered updating approach.
For reference, the following test would fail:
```kotlin
@Test
fun testFlatMapDynamicParentIsCreatedLazily() {
    val staticParent = mutableProvider(0)
    lateinit var dynamicParent: MutableProvider<String>
    val flatMapped = staticParent.flatMapMutable {
        dynamicParent = mutableProvider("a")
        dynamicParent
    }
    flatMapped.set("b")
    assertEquals("b", dynamicParent.get())
}
```
This is because, due to the lazy evaluation of the transform function, the mutableProvider("a") is actually created after .set("b"), which prevents "b" from being stored in the provider, as "a" is technically newer. This is of course unwanted behavior, but I could not find a way to solve this without switching away from the sequence-numbered approach.
Approaches considered:
1. re-emit on set: does not work as it breaks ordering
2. A pre-update loop that is run on .set that resolves the dynamic parent, then creates the DeferredValue (with a larger seqNo) after. This does not work because there may be a race that un-initializes the dynamic parent inbetween the pre-update and update calls. This would require locking over the entire provider graph which is not possible with the current architecture and what sequence-numbering is trying to avoid in the first place.
3. Add additional deferred values: InitialDirect, InitialLazy which always have seqNo -1. Then only use the regular Direct and Lazy on .set calls. This would work for most use-cases, but it does not solve the problem. An additional set call inside the transform function would break this again.